fix: move remote image fetching to sourceNodes
- This fixes LS-1436
- Solves issue where remote images won't resolve with Gatsby 4
