+++ packages/gatsby-source-nacelle/src/source-nodes/index.js
@@ -1,3 +1,9 @@
+const {
+  replaceKey,
+  cacheIsInvalid,
+  hasBeenIndexedSinceLastBuild,
+  createRemoteImageFileNode
+} = require('../utils');
 /**
  * Creates Gatsby nodes from Nacelle data
  * @param {Object} config - configuration object
@@ -19,12 +25,6 @@ module.exports = async function ({
   ],
   uniqueIdProperty = 'remoteId'
 }) {
-  const {
-    replaceKey,
-    cacheIsInvalid,
-    hasBeenIndexedSinceLastBuild
-  } = require('../utils');
-
   const { actions, createContentDigest, cache } = gatsbyApi;
   const { createNode, touchNode } = actions;
 
@@ -45,6 +45,16 @@ module.exports = async function ({
   const dataTypeUpper =
     dataType.charAt(0).toUpperCase() + dataType.substring(1, dataType.length);
 
+  // detect if users have opted into gatsby image
+  let useGatsbyImage = false;
+  try {
+    // the user can opt into using Gatsby Image by installing `gatsby-source-filesystem`
+    require('gatsby-source-filesystem');
+    useGatsbyImage = true;
+  } catch (err) {
+    // do nothing because useGatsbyImage is already false
+  }
+
   try {
     console.info(`[gatsby-source-nacelle] fetching ${dataTypeLower}`);
 
@@ -59,7 +69,7 @@ module.exports = async function ({
       : replaceKey(data, keyMappings);
 
     if (Array.isArray(formattedData)) {
-      formattedData.forEach((entry) => {
+      formattedData.forEach(async (entry) => {
         if (
           hasBeenIndexedSinceLastBuild(entry, lastFetched) ||
           cacheIsInvalid(lastFetched, pluginOptions)
@@ -75,6 +85,9 @@ module.exports = async function ({
           };
 
           const node = Object.assign({}, entry, nodeMeta);
+          if (useGatsbyImage) {
+            await fetchRemoteImageNodes(dataType, node, gatsbyApi);
+          }
           createNode(node);
 
           newNodeCount += 1;
@@ -97,6 +110,9 @@ module.exports = async function ({
       };
 
       const node = Object.assign({}, formattedData, nodeMeta);
+      if (useGatsbyImage) {
+        await fetchRemoteImageNodes(dataType, node, gatsbyApi);
+      }
       createNode(node);
     }
 
@@ -117,3 +133,35 @@ module.exports = async function ({
     );
   }
 };
+
+/**
+ * helper for fetching images based on the content type.
+ * @param {string} dataType - kind of node data
+ * @param {Object} node - the new node to be created
+ * @param {Object} gatsbyApi - Functions provided by `Gatsby for managing nodes`
+ * @param {Object} gatsbyApi.actions
+ * @param {function} gatsbyApi.getCache
+ * @param {function} gatsbyApi.createNodeId
+ */
+async function fetchRemoteImageNodes(dataType, node, gatsbyApi) {
+  const isImage = (nodeMediaEntry) =>
+    nodeMediaEntry &&
+    nodeMediaEntry.type &&
+    nodeMediaEntry.type.startsWith('image');
+  if (dataType === 'product') {
+    await createRemoteImageFileNode(
+      node,
+      ['featuredMedia', 'media'],
+      gatsbyApi,
+      { isImage }
+    );
+  } else if (dataType === 'collection') {
+    await createRemoteImageFileNode(node, 'featuredMedia', gatsbyApi, {
+      isImage
+    });
+  } else if (dataType === 'content') {
+    await createRemoteImageFileNode(node, 'featuredMedia', gatsbyApi, {
+      isImage
+    });
+  }
+}
